% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarySegmentation.R
\name{binarySegmentation_interface}
\alias{binarySegmentation_interface}
\alias{binarySegmentation_algo}
\title{Binary Segmentation}
\usage{
binarySegmentation_algo(data)
}
\arguments{
\item{data}{The input numeric matrix.}
}
\value{
A list containing eigenvalues, eigenvectors, and principal components.
}
\description{
This function calls a C++ function to perform Binary Segmentation on the input data.
}
\examples{
set.seed(123)
data <- c(rnorm(50, mean = 2), rnorm(50, mean = 8))
binarySegmentation_algo(data)
}
