% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicSegmentation.R
\name{dynamicSegmentation_interface}
\alias{dynamicSegmentation_interface}
\alias{dynamicSegmentation_algo}
\title{Dynamic Segmentation Algorithm}
\usage{
dynamicSegmentation_algo(data, threshold)
}
\arguments{
\item{data}{The input numeric vector.}

\item{threshold}{The threshold value.}
}
\value{
An integer vector of changepoint indices.
}
\description{
This function performs dynamic segmentation on the input data.
}
\examples{
set.seed(123)
data <- rnorm(100)
threshold = 0.5
dynamicSegmentation_algo(data, threshold)
}
